+++ app/models/artefact.rb
@@ -118,6 +118,8 @@ class Artefact
                                   "written_statement"],
   }.freeze
 
+  RETIRED_FORMATS = ["video"]
+
   FORMATS = FORMATS_BY_DEFAULT_OWNING_APP.values.flatten
 
   def self.default_app_for_format(format)